Support subtype arguments
The existing macro fails when called when the method arguments are
subtypes of the method parameters. For example, suppose we have:

class Foo
class Bar extends Foo

With the previous version of the library, the following doesn't work:

object Test {
  def apply(foo: Foo): Int = 1
}
Thunk(Test(new Bar))

because there is no method with name apply and Bar type (the type of the
foo parameter is foo). To fix this, we need to look up all of the
methods in the class and check if the method has the correct name and
compatible arguments. The arguments are compatible when the method
argument list has the same number of parameters as input arguments and
for each of the arguments, the input argument type is a subtype of the
corresponding method parameter type. Unlike in java, we can do this at
compile time, which should improve the performance.