Improve ChildFirstClassLoader
I reworked the ChildFirstClassLoader to make it a bit more user
friendly. I simplified the primary public constructor to just take three
arguments: the URLs of the class path for the URLClassLoader, a function
from (className: String) => RequiredClassLoader and the parent
classloader to use. Since most of the time the parent will just be the
current Thread.getContextClassLoader and we always want to try the
URLClassLoader first, I added a one argument constructor that just takes
the URL array.

Bonus: improve javadocs and move private method to bottom of the class