Add ClassLoaders
This commit adds some java utilities for reflectively creating object
instances and invoking methods. This mimics some of the functionality
available in the Thunk object in the scala project. It abstracts away
the java reflection library so that it's relatively easy to use the
reflection api without a massive amount of boilerplate.