Add ArgParser mini-library
I am going to add a class that wraps a main class and restarts the class
with a new classloader whenever class or resource files change. Since
this will be run from the command line, I want there to be good support
for passing in command line arguments. I didn't want to add an existing
argument parsing library because I intend for the core swoval libraries
to be self contained, or possibly contain dependencies to other swoval
libraries.

The argument parsing mini-library has two main features:
1) It groups arguments by key name. The key names are specified when the
   parser is initialized. An argument may have multiple names but, if it
   does, it's up to the application to extract the values from both
   names.

2) It provides a helper to print a useful help or startup message